requirements.txt
Flask==3.0.1
psycopg2-binary==2.9.10
dotenv==0.9.9